reorg: send-footer
