--- ui/app/pages/send/send-footer/send-footer.selectors.js
@@ -1,5 +1,0 @@
-import { getSendErrors } from '../send.selectors'
-
-export function isSendFormInError (state) {
-  return Object.values(getSendErrors(state)).some((n) => n)
-}
